+++ packages/devtools_shared/lib/src/deeplink/app_link_settings.dart
@@ -12,10 +12,14 @@ class AppLinkSettings {
   const AppLinkSettings._(this.applicationId, this.deeplinks);
 
   factory AppLinkSettings.fromJson(String json) {
-    final jsonObject = jsonDecode(json);
+    final jsonObject = jsonDecode(json) as Map;
+    final {
+      _kApplicationIdKey: String applicationId,
+      _kDeeplinksKey: List<Object?> deepLinks,
+    } = jsonObject;
     return AppLinkSettings._(
-      jsonObject[_kApplicationIdKey] as String,
-      (jsonObject[_kDeeplinksKey] as List<dynamic>)
+      applicationId,
+      deepLinks
           .cast<Map<String, dynamic>>()
           .map<AndroidDeeplink>(AndroidDeeplink._fromJsonObject)
           .toList(),