+++ tool/lib/commands/serve.dart
@@ -13,7 +13,12 @@ import 'package:path/path.dart' as path;
 import 'shared.dart';
 
 const _buildAppFlag = 'build-app';
+const _debugServerFlag = 'debug-server';
+
+// TODO(https://github.com/flutter/devtools/issues/7232): Consider using
+// AllowAnythingParser instead of manually passing these args through.
 const _machineFlag = 'machine';
+const _dtdUriFlag = 'dtd-uri';
 const _allowEmbeddingFlag = 'allow-embedding';
 
 /// This command builds DevTools in release mode by running the
@@ -25,6 +30,12 @@ const _allowEmbeddingFlag = 'allow-embedding';
 /// are ignored if '--no-build-app' is present in the list of arguments passed
 /// to this command. All of the following commands are passed along to the
 /// `devtools_tool build` command.
+///
+/// If the [_debugServerFlag] argument is present, the DevTools server will be
+/// started with the `--observe` flag. This will allow you to debug and profile
+/// the server with a local VM service connection. By default, this will set
+/// `--pause-isolates-on-start` and `--pause-isolates-on-unhandled-exceptions`
+/// for the DevTools server VM service connection.
 ///
 /// If the [BuildCommandArgs.useFlutterFromPath] argument is present, the
 /// Flutter SDK will not be updated to the latest Flutter candidate before
@@ -57,6 +68,12 @@ class ServeCommand extends Command {
             'Whether to build the DevTools web app before starting the DevTools'
             ' server.',
       )
+      ..addFlag(
+        _debugServerFlag,
+        negatable: false,
+        defaultsTo: false,
+        help: 'Enable debugging for the DevTools server.',
+      )
       ..addUpdateFlutterFlag()
       ..addUpdatePerfettoFlag()
       ..addPubGetFlag()
@@ -66,6 +83,10 @@ class ServeCommand extends Command {
         _machineFlag,
         negatable: false,
         help: 'Sets output format to JSON for consumption in tools.',
+      )
+      ..addOption(
+        _dtdUriFlag,
+        help: 'Sets the dtd uri when starting the devtools server',
       )
       ..addFlag(
         _allowEmbeddingFlag,
@@ -83,10 +104,17 @@ class ServeCommand extends Command {
 
   @override
   Future run() async {
+    logStatus(
+      'WARNING: if you have local changes in packages/devtools_shared, you will'
+      ' need to add a path dependency override in sdk/pkg/dds/pubspec.yaml in'
+      ' order for these changes to be picked up.',
+    );
+
     final repo = DevToolsRepo.getInstance();
     final processManager = ProcessManager();
 
     final buildApp = argResults![_buildAppFlag];
+    final debugServer = argResults![_debugServerFlag];
     final updateFlutter = argResults![BuildCommandArgs.updateFlutter.flagName];
     final updatePerfetto =
         argResults![BuildCommandArgs.updatePerfetto.flagName];
@@ -94,12 +122,14 @@ class ServeCommand extends Command {
     final devToolsAppBuildMode =
         argResults![BuildCommandArgs.buildMode.flagName];
 
+    // Any flag that we aren't removing here is intended to be passed through.
     final remainingArguments = List.of(argResults!.arguments)
       ..remove(BuildCommandArgs.updateFlutter.asArg())
       ..remove(BuildCommandArgs.updateFlutter.asArg(negated: true))
       ..remove(BuildCommandArgs.updatePerfetto.asArg())
       ..remove(valueAsArg(_buildAppFlag))
       ..remove(valueAsArg(_buildAppFlag, negated: true))
+      ..remove(valueAsArg(_debugServerFlag))
       ..remove(BuildCommandArgs.pubGet.asArg())
       ..remove(BuildCommandArgs.pubGet.asArg(negated: true))
       ..removeWhere(
@@ -163,6 +193,10 @@ class ServeCommand extends Command {
     await processManager.runProcess(
       CliCommand.dart(
         [
+          if (debugServer) ...[
+            'run',
+            '--observe=0',
+          ],
           serveLocalScriptPath,
           '--devtools-build=$devToolsBuildLocation',
           // Pass any args that were provided to our script along. This allows IDEs