+++ packages/devtools_extensions/CHANGELOG.md
@@ -4,6 +4,9 @@
 simulated DevTools environment.
 * Add help buttons to the VM Service and DTD connection text fields in the
 simulated DevTools environment.
+* Bump `package:vm_service` dependency to ^14.0.0.
+* Refactor `example` directory to support more package examples.
+* Add an example of providing an extension from a pure Dart package.
 
 ## 0.0.13
 * Bump `package:web` to `^0.4.1`.